+++ config/openapi.php
@@ -260,7 +260,9 @@ return [
         // Requires `composer require spatie/laravel-query-builder`. Uncomment to enable:
         // \Radiergummi\OpenApi\Plugins\QueryBuilder\QueryBuilderPlugin::class,
 
-        // Requires `composer require league/fractal`. Uncomment to enable:
+        // Requires either `composer require league/fractal` or
+        // `composer require spatie/laravel-fractal` (which depends on league/fractal).
+        // Uncomment to enable:
         // \Radiergummi\OpenApi\Plugins\Fractal\FractalPlugin::class,
     ],